Remove redundant service
